Pack & unpack MyUnfolder using pickle
Some nasty hacking in there - for some reason, although all the individual attributes _can_ be pickled, the whole object cannot.
So now I have to pickle each object individually (with a list of their names first to keep track of them),
then when unpickleing we must create a blank MyUnfolder, then fill it with the objects.